refactor: move HTTP utilities into pkg/util/http
