+++ pkg/debuginfo/store_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/grafana/pyroscope/api/gen/proto/go/debuginfo/v1alpha1/debuginfov1alpha1connect"
 	"github.com/grafana/pyroscope/v2/pkg/objstore/providers/memory"
 	"github.com/grafana/pyroscope/v2/pkg/tenant"
-	"github.com/grafana/pyroscope/v2/pkg/util"
+	httputil "github.com/grafana/pyroscope/v2/pkg/util/http"
 )
 
 func newTestStore(t *testing.T, cfg Config) (*Store, *memory.InMemBucket) {
@@ -521,7 +521,7 @@ func startTestServer(t *testing.T, store *Store) testServer {
 	)
 	router.Handle(
 		"/debuginfo.v1alpha1.DebuginfoService/Upload/{gnu_build_id}",
-		util.AuthenticateUser(true).Wrap(store.UploadHTTPHandler()),
+		httputil.AuthenticateUser(true).Wrap(store.UploadHTTPHandler()),
 	).Methods("POST")
 
 	srv := httptest.NewServer(router)
