+++ pkg/util/http/http_test.go
@@ -1,4 +1,4 @@
-package util
+package http
 
 import (
 	"bytes"
